refactor(integrations): split booking.rs ota_xml into submodule file
Extract the self-contained `ota_xml` module (~1480 lines) out of the
4295-line `booking.rs` churn hotspot into its own `booking/ota_xml.rs`
file, converting `booking.rs` to a `booking/mod.rs` directory module.

Pure, behavior-preserving move: no logic, signatures, or public paths
change. `booking::ota_xml::*` resolves identically (lib.rs re-exports and
api-server's `integrations::booking::ota_xml` import are unchanged); the
module's `super::`/`crate::booking::` references and its own `#[cfg(test)]`
tests move with it. This reduces the churn surface of the main file by
~34% (4295 -> 2818 lines) so future Booking.com OTA retry work (cf. PR
#1294, +404/-29) lands in a smaller, more reviewable file.

Cargo.lock picks up a pre-existing `rust_decimal` entry for `integrations`
(already declared in its Cargo.toml on dev) that the stale lock was missing.

Verified: cargo check/clippy (-D warnings)/test -p integrations all green
(67 booking tests pass).
